# Improve perf of T::Enum when not using enable_legacy_t_enum_migration_mode